pyproject.toml
[project]
name = "hybrid-sentinel"
version = "0.1.0"
description = "Real-time AI transaction anomaly detection and agentic investigation for P2P payment routing"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Hrachov Oleksandr" },
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "bytewax>=0.21.0",
    "river>=0.21.0",
]

[project.optional-dependencies]
agent = [
    "langgraph>=0.2.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hybrid_sentinel"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "integration: end-to-end tests requiring the full Bytewax pipeline (may be slow)",
]

[tool.ruff]
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
mypy_path = "src"
strict = false
warn_return_any = true
warn_unused_configs = true
